Add benefit freshness pipeline (monitoring, review, publishing)
Detect credit-card benefit changes from authoritative issuer sources, stage
them as evidence-backed proposals, review them in a new admin surface, and
publish approved changes transactionally with full audit history and rollback.
Scraping failures or model errors can never corrupt production data.

- ~72 new core source/doc files (114 new incl. colocated tests, fixtures, pgTAP)
- 2 new additive migrations: pipeline schema + service-role publish/rollback RPCs
- 3 idempotent reproducibility fixes to pre-existing migrations (fresh-DB only)
- Daily cron /api/cron/monitor-sources; allowlisted outbound fetch; Anthropic
  extraction behind a provider-agnostic, env-configured interface
- Admin surface gated by ADMIN_EMAILS; all sources ship disabled; no auto-publish

Verified: 165/165 JS tests, 124/124 pgTAP (twice across clean db resets),
clean tsc/lint/build.

+++ supabase/migrations/20260212120000_benefits_mvp_schema.sql
@@ -431,7 +431,9 @@ select
   seeded_benefits.display_name,
   seeded_benefits.category,
   seeded_benefits.cadence,
-  seeded_benefits.value_cents,
+  -- All seeded rows carry NULL here, so the VALUES column resolves to text;
+  -- cast explicitly so the insert matches benefits.value_cents (integer).
+  seeded_benefits.value_cents::integer,
   seeded_benefits.requires_enrollment,
   seeded_benefits.requires_selection,
   seeded_benefits.selection_type,
